requirements.txt
requests
jinja2
docopt
python-dateutil
emoji